% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbox_dists}
\alias{dbox_dists}
\title{A set of random distributions for visualization tests, genereated by
densitybox::test_dists().
Distributions include: normal, right- and left-skewed, uniform, bimodal,
absolute value sine, and a "weird" distribution of concatenated
chi-squared's with random parameters.}
\format{An object of class \code{data.frame} with 10000 rows and 7 columns.}
\usage{
dbox_dists
}
\description{
A set of random distributions for visualization tests, genereated by
densitybox::test_dists().
Distributions include: normal, right- and left-skewed, uniform, bimodal,
absolute value sine, and a "weird" distribution of concatenated
chi-squared's with random parameters.
}
\keyword{datasets}
