% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guilford}
\alias{guilford}
\title{Guilford County, NC 2017 Income Data}
\format{A data frame of 3,352 observations and 2 variables.
\describe{
}}
\usage{
guilford
}
\description{
3,352 observations plus survey weights of 2017 Guilford County, NC personal
income. Source: American Community Survey, via IPUMS USA.
}
\keyword{datasets}
