% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitybox.R
\name{dbox}
\alias{dbox}
\title{Combined density/box plots.}
\usage{
dbox(x, color = "black", alpha = 0.2, linewt = 1.1,
  linewt_box = 0.8, normal = FALSE, color_norm = "black",
  linewt_norm = 1.1)
}
\arguments{
\item{x}{A continuous numeric vector.}

\item{color}{Fill and line color of both density and box plots.}

\item{alpha}{Fill transparency of both density and box plots.}

\item{linewt}{Weight of density plot outline.}

\item{linewt_box}{Weight of boxplot lines.}

\item{normal}{Logical. Include a simulated normal/Gaussian density line plot with the same parameters as x? Defaults to FALSE.}

\item{color_norm}{Line color of the optional simulated normal plot.}

\item{linewt_norm}{Wieght of the optional simulated normal plot.}
}
\description{
Combined density/box plots.
}
\examples{
ndensity <- rnorm(5000)
dbox(ndensity)
}
