% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbox.R
\name{dbox}
\alias{dbox}
\title{Combined density/box plots.}
\usage{
dbox(x, w = NULL, coef = 1.5, color = "black", alpha = 0.2,
  lwt = 1.1, ltype = "solid", fill = TRUE, normal = FALSE,
  color_norm = "black", lwt_norm = 1.1, lytpe_norm = "twodash")
}
\arguments{
\item{x}{A continuous numeric vector.}

\item{w}{A vector of weights. Must be either NULL (the default) or the
same length as <x>.}

\item{coef}{Defines the length of boxplot whiskers & outliers. Defaults
to the typical stat_boxplot value of 1.5.}

\item{color}{Fill and line color of both density and box plots.}

\item{alpha}{Fill transparency of both density and box plots.}

\item{lwt}{Weight of density plot outline.}

\item{ltype}{Line type (pattern) of density plot outline. Default is solid.
For other options, see ?linetype.}

\item{fill}{Logical. Fill the density plot? Gets color from <color> and
alpha from <alpha>. Defaults to TRUE}

\item{normal}{Logical. Include a simulated normal/Gaussian density
plot? Gets parameters from <x> Defaults to FALSE.}

\item{color_norm}{Line color of the optional simulated normal plot.}

\item{lwt_norm}{Weight of the optional simulated normal plot.}

\item{ltype_norm}{Line type simulated normal. Default is twodash.}
}
\description{
Combined density/box plots.
}
\examples{
ndensity <- rnorm(5000)
dbox(ndensity)

rskew <- 1.5^rnorm(5000) * 5
dbox(rskew, color = "orange2", fill = FALSE,
     normal = TRUE, color_norm = "steelblue2")
}
