% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dists.R
\name{test_dists}
\alias{test_dists}
\title{Generate a dataframe of several random distributions for visualization tests.
Distributions include: normal, right- and left-skewed, uniform, bimodal,
absolute value sine, and a "weird" distribution of concatenated
chi-squared's with random parameters.}
\usage{
test_dists(n)
}
\arguments{
\item{n}{Number of observations.}
}
\description{
Generate a dataframe of several random distributions for visualization tests.
Distributions include: normal, right- and left-skewed, uniform, bimodal,
absolute value sine, and a "weird" distribution of concatenated
chi-squared's with random parameters.
}
\examples{
testing <- test_dists(5000)
}
